requirements.txt
# Requirements für die Bildverarbeitung und Maskenextraktion
opencv-python>=4.12.0
pillow>=11.3.0
numpy>=2.2.6


# Für LRP-Analyse und Excel-Export
pandas>=2.2.0
openpyxl>=3.1.2

# Für Linear Probing (Logistische Regression, SGD)
scikit-learn>=1.5.0


# Currently given requirements (from maskdino-arm conda environment):
absl-py==2.3.1
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
black==21.4b2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
cloudpickle==3.1.1
contourpy==1.3.2
cycler==0.12.1
cython==3.1.3
detectron2==0.6
filelock==3.19.1
fonttools==4.59.2
fsspec==2025.7.0
future==1.0.0
fvcore==0.1.5.post20221221
grpcio==1.74.0
h5py==3.14.0
hf-xet==1.1.9
huggingface-hub==0.34.4
hydra-core==1.3.2
idna==3.10
imageio==2.37.0
iopath==0.1.9
jinja2==3.1.6
kiwisolver==1.4.9
lazy-loader==0.4
markdown==3.8.2
markupsafe==3.0.2
matplotlib==3.10.5
mpmath==1.3.0
mypy-extensions==1.1.0
networkx==3.4.2
numpy==2.2.6
omegaconf==2.3.0
opencv-python==4.12.0.88
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
portalocker==3.2.0
protobuf==6.32.0
pycocotools==2.0.10
pydot==4.0.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pyyaml==6.0.2
regex==2025.7.34
requests==2.32.5
safetensors==0.6.2
scikit-image==0.25.2
scipy==1.15.3
shapely==2.1.1
six==1.17.0
submitit==1.5.3
sympy==1.14.0
tabulate==0.9.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
termcolor==3.1.0
tifffile==2025.5.10
timm==1.0.19
toml==0.10.2
torch==2.8.0
torchaudio==2.8.0
torchvision==0.23.0
tqdm==4.67.1
typing-extensions==4.15.0
urllib3==2.5.0
werkzeug==3.1.3
yacs==0.1.8